+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/Rules/Deleg.hs
@@ -12,3 +12,6 @@ import qualified Cardano.Ledger.Shelley.Rules as Shelley
 type instance EraRuleFailure "DELEG" AllegraEra = Shelley.ShelleyDelegPredFailure AllegraEra
 
 instance InjectRuleFailure "DELEG" Shelley.ShelleyDelegPredFailure AllegraEra
+
+instance InjectRuleFailure "DELEG" Shelley.AccountAlreadyRegistered AllegraEra where
+  injectFailure = Shelley.DelegAccountAlreadyRegistered