+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Rules/Deleg.hs
@@ -12,3 +12,6 @@ import qualified Cardano.Ledger.Shelley.Rules as Shelley
 type instance EraRuleFailure "DELEG" AlonzoEra = Shelley.ShelleyDelegPredFailure AlonzoEra
 
 instance InjectRuleFailure "DELEG" Shelley.ShelleyDelegPredFailure AlonzoEra
+
+instance InjectRuleFailure "DELEG" Shelley.AccountAlreadyRegistered AlonzoEra where
+  injectFailure = Shelley.DelegAccountAlreadyRegistered