+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Deleg.hs
@@ -1,13 +1,17 @@
 {-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
@@ -16,6 +20,10 @@ module Cardano.Ledger.Shelley.Rules.Deleg (
   DelegEnv (..),
   ShelleyDelegPredFailure (..),
   ShelleyDelegEvent (..),
+
+  -- * Predicate failures
+  AccountAlreadyRegistered (..),
+  checkAccountAlreadyRegistered,
 ) where
 
 import Cardano.Ledger.BaseTypes (
@@ -44,6 +52,7 @@ import Cardano.Ledger.Coin (
 import Cardano.Ledger.Compactible (fromCompact)
 import Cardano.Ledger.Credential (Credential, Ptr)
 import Cardano.Ledger.Hashes (GenDelegPair (..), GenDelegs (..))
+import Cardano.Ledger.Rules.ValidationMode ((?!.))
 import Cardano.Ledger.Shelley.Core
 import Cardano.Ledger.Shelley.Era (DELEG, ShelleyEra, hardforkAlonzoAllowMIRTransfer)
 import Cardano.Ledger.Shelley.LedgerState (availableAfterMIR)
@@ -70,10 +79,38 @@ import Data.Word (Word8)
 import GHC.Generics (Generic)
 import Lens.Micro
 
+-- | Predicate failure that is reported upon an attempt to register an account with a staking
+-- credential that is already registered.
+newtype AccountAlreadyRegistered era
+  = AccountAlreadyRegistered (Credential Staking)
+  deriving stock (Show, Generic)
+  deriving newtype (Eq, Ord, NFData, EncCBOR, DecCBOR)
+
+instance InjectRuleFailure "DELEG" AccountAlreadyRegistered ShelleyEra where
+  injectFailure = DelegAccountAlreadyRegistered
+
+-- | `AccountAlreadyRegistered` predicate failure check.
+--
+-- In Shelley spec defined in:
+--
+-- === Figure 23: Delegation Inference Rules (8)
+--
+-- @
+-- (hk ∉ dom (rewards ds))
+-- @
+checkAccountAlreadyRegistered ::
+  forall rule sts era.
+  ( InjectRuleFailure rule AccountAlreadyRegistered era
+  , PredicateFailure sts ~ EraRuleFailure rule era
+  , EraAccounts era
+  ) =>
+  Accounts era -> Credential Staking -> Rule sts 'Transition ()
+checkAccountAlreadyRegistered accounts cred =
+  not (isAccountRegistered cred accounts) ?!. AccountAlreadyRegistered cred
+
 data DelegEnv era = DelegEnv
   { slotNo :: SlotNo
   , deCurEpochNo :: EpochNo
-  -- ^ Lazy on purpose, because not all certificates need to know the current EpochNo
   , ptr_ :: Ptr
   , deChainAccountState :: ChainAccountState
   , ppDE :: PParams era -- The protocol parameters are only used for the HardFork mechanism
@@ -87,8 +124,7 @@ deriving instance Eq (PParams era) => Eq (DelegEnv era)
 instance NFData (PParams era) => NFData (DelegEnv era)
 
 data ShelleyDelegPredFailure era
-  = StakeKeyAlreadyRegisteredDELEG
-      (Credential Staking) -- Credential which is already registered
+  = DelegAccountAlreadyRegistered (AccountAlreadyRegistered era)
   | StakeKeyNotRegisteredDELEG
       (Credential Staking) -- Credential which is not registered
   | StakeKeyNonZeroAccountBalanceDELEG
@@ -136,6 +172,8 @@ instance
   , ShelleyEraAccounts era
   , ShelleyEraTxCert era
   , AtMostEra "Babbage" era
+  , EraRuleFailure "DELEG" era ~ ShelleyDelegPredFailure era
+  , InjectRuleFailure "DELEG" AccountAlreadyRegistered era
   ) =>
   STS (DELEG era)
   where
@@ -152,7 +190,7 @@ instance NFData (ShelleyDelegPredFailure era)
 
 instance Era era => EncCBOR (ShelleyDelegPredFailure era) where
   encCBOR = \case
-    StakeKeyAlreadyRegisteredDELEG cred ->
+    DelegAccountAlreadyRegistered cred ->
       encodeListLen 2 <> encCBOR (0 :: Word8) <> encCBOR cred
     StakeKeyNotRegisteredDELEG cred ->
       encodeListLen 2 <> encCBOR (1 :: Word8) <> encCBOR cred
@@ -202,7 +240,7 @@ instance
     \case
       0 -> do
         kh <- decCBOR
-        pure (2, StakeKeyAlreadyRegisteredDELEG kh)
+        pure (2, DelegAccountAlreadyRegistered kh)
       1 -> do
         kh <- decCBOR
         pure (2, StakeKeyNotRegisteredDELEG kh)
@@ -255,6 +293,8 @@ delegationTransition ::
   , ShelleyEraTxCert era
   , EraPParams era
   , AtMostEra "Babbage" era
+  , EraRuleFailure "DELEG" era ~ ShelleyDelegPredFailure era
+  , InjectRuleFailure "DELEG" AccountAlreadyRegistered era
   ) =>
   TransitionRule (DELEG era)
 delegationTransition = do
@@ -263,8 +303,7 @@ delegationTransition = do
       ds = certState ^. certDStateL
   case c of
     RegTxCert cred -> do
-      -- (hk ∉ dom (rewards ds))
-      not (isAccountRegistered cred (ds ^. accountsL)) ?! StakeKeyAlreadyRegisteredDELEG cred
+      checkAccountAlreadyRegistered (ds ^. accountsL) cred
       let compactDeposit = compactCoinOrError (pp ^. ppKeyDepositL)
       pure $ certState & certDStateL . accountsL %~ registerShelleyAccount cred ptr compactDeposit Nothing
     UnRegTxCert cred -> do
@@ -390,6 +429,8 @@ checkSlotNotTooLate ::
   , ShelleyEraTxCert era
   , EraPParams era
   , AtMostEra "Babbage" era
+  , EraRuleFailure "DELEG" era ~ ShelleyDelegPredFailure era
+  , InjectRuleFailure "DELEG" AccountAlreadyRegistered era
   ) =>
   SlotNo ->
   EpochNo ->