+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Arbitrary.hs
@@ -52,6 +52,7 @@ import Cardano.Ledger.Shelley.Rewards (
   StakeShare (..),
  )
 import Cardano.Ledger.Shelley.Rules (
+  AccountAlreadyRegistered (..),
   ShelleyBbodyPredFailure,
   ShelleyDelegPredFailure,
   ShelleyDelegsPredFailure,
@@ -676,6 +677,8 @@ instance
   where
   arbitrary = genericArbitraryU
   shrink = genericShrink
+
+deriving instance Arbitrary (AccountAlreadyRegistered era)
 
 instance
   ( Era era