+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/TreeDiff.hs
@@ -198,6 +198,8 @@ instance ToExpr (ShelleyPoolPredFailure era)
 -- Rules/Deleg
 instance ToExpr (ShelleyDelegPredFailure era)
 
+instance ToExpr (AccountAlreadyRegistered era)
+
 -- Rules/Utxow
 instance
   ( Era era