+++ eras/conway/impl/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Version history for `cardano-ledger-conway`
 
-## 1.23.1.0
+## 1.24.0.0
 
+* Replace `StakeKeyRegisteredDELEG` constructor with `DelegAccountAlreadyRegistered` in `ConwayDelegPredFailure`, which wraps the new `AccountAlreadyRegistered` type instead of `Credential Staking`
 * Add `EncCBOR`, `ToCBOR` for `Block`
 * Add `DecCBOR` instances for `Annotator Block`